@@ -6,18 +6,15 @@ import Box from '@mui/material/Box'
 import { Outlet } from "react-router-dom"
 import logo from '../../assets/images/logo.svg'
 import Button from '@mui/material/Button'
-
+import TextSnippetIcon from '@mui/icons-material/TextSnippet';
 function Home() {
   return (
     <Box>
-      <AppBar position="static" color='transparent'>
+      <AppBar position="static" color='default'>
         <Container maxWidth="xl">
           <Toolbar >
-            <img src={logo} alt="logo" style={{ width: 60 }} />
             <Typography
-              variant='h1'
-              component="div"
-              color="secondary"
+              variant='h5'
               sx={{ display: {xs: 'none', lg: 'block', xl: 'block'} }}
             >
               Ngrancang Board
@@ -25,15 +22,16 @@ function Home() {
             <Box display="flex" justifyContent="end" flexGrow={1}>
               <Button
                 sx={{mx: 2}}
-                variant="contained"
-                color="primary"
+                href="/login"
+                variant='contained'
               >
-                  Login
+                Login
               </Button>
               <Button
                 sx={{mx: 2}}
                 variant="contained"
                 color='secondary'
+                href="/register"
               >
                   Daftar
               </Button>
